pyproject.toml
[tool.poetry]
name = "dlboost"
version = "0.1.0"
description = ""
authors = ["Chunxu Guo <chunxu@wustl.edu>"]

[tool.poetry.dependencies]
python = "^3.11"
torch = {version = "^2.2.0+cu121", source = "torch_cuda"}
torchvision = {version = "^0.17.0+cu121", source = "torch_cuda"}
scipy = "^1.12.0"
numpy = "^1.26.4"
einops = "^0.7.0"
tqdm = "^4.66.1"
toolz = "^0.12.1"
h5py = "^3.10.0"
nibabel = "^5.2.0"
pydicom = "^2.4.4"
zarr = "^2.16.1"
monai = "^1.3.0"
lightning = {extras = ["pytorch-extra"], version = "^2.1.4"}
torchopt = "^0.7.3"
wandb = "^0.16.3"
napari = "^0.4.19"
jupyterlab = "^4.1.0"
mrboost = {path = "../mrboost", develop = true}
torchkbnufft = "^1.4.0"
matplotlib = "^3.8.2"
xarray = {extras = ["parallel"], version = "^2024.1.1"}
nvitop = "^1.3.2"
pytorch-minimize = "^0.0.2"
jaxlib = {version = "^0.4.24", source = "jaxlib"}
jax = {version = "^0.4.24", source = "jax"}
cufinufft = "^2.2.0"
mri-nufft = "^0.4.0"
pytorch-finufft = {git = "https://github.com/flatironinstitute/pytorch-finufft.git", branch = "main"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cuda12"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_releases.html"
priority = "explicit"

[[tool.poetry.source]]
name = "jaxlib"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "explicit"


